refactor: moved `areLabelsEqual` function into `helpers` file
